Fixed the tests correctly
